Lemmatization, tokenization, and padding
The data was lemmatized to reduce the words to their root words or
'lemma.' This reduces the dimensionality of the data.

Then, the data was tokenized because words of subwords need to
be treated as distinct entities in order to understand their
relationship.

The data was padded, meaning to add additional values to data to make
it fit a particular length. All data model input data should have the
same length.