Max sequence length, Vectorize, Split data, LSTM, Visualize
Chose max length to standardize input sequences so they can be fed
into the model.

Vectorized the data which means putting it into a numerical format so
it's suitable for machine learning.

Split the data so as to have a set to train the model on and another to
test the accuracy of said model.

Long Short Term Memory LSTM is the model that was used to conduct the
sentinment analysis.

Visualized the accuracy of the model.